Refactor DotNetAliases: Extract methods into a separate partial class
