+++ src/Cake.Common/Tools/DotNet/DotNetAliases.Execute.cs
@@ -1,0 +1,101 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using Cake.Common.Tools.DotNet.Execute;
+using Cake.Core;
+using Cake.Core.Annotations;
+using Cake.Core.IO;
+
+namespace Cake.Common.Tools.DotNet
+{
+    /// <summary>
+    /// <para>Contains functionality related to <see href="https://github.com/dotnet/cli">.NET CLI</see>.</para>
+    /// <para>
+    /// In order to use the commands for this alias, the .NET CLI tools will need to be installed on the machine where
+    /// the Cake script is being executed.  See this <see href="https://www.microsoft.com/net/core">page</see> for information
+    /// on how to install.
+    /// </para>
+    /// </summary>
+    public static partial class DotNetAliases
+    {
+        /// <summary>
+        /// Execute an assembly.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="assemblyPath">The assembly path.</param>
+        /// <example>
+        /// <code>
+        /// DotNetExecute("./bin/Debug/app.dll");
+        /// </code>
+        /// </example>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Execute")]
+        [CakeNamespaceImport("Cake.Common.Tools.DotNet.Execute")]
+        public static void DotNetExecute(this ICakeContext context, FilePath assemblyPath)
+        {
+            context.DotNetExecute(assemblyPath, null);
+        }
+
+        /// <summary>
+        /// Execute an assembly with arguments in the specific path.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="assemblyPath">The assembly path.</param>
+        /// <param name="arguments">The arguments.</param>
+        /// <example>
+        /// <code>
+        /// DotNetExecute("./bin/Debug/app.dll", "--arg");
+        /// </code>
+        /// </example>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Execute")]
+        [CakeNamespaceImport("Cake.Common.Tools.DotNet.Execute")]
+        public static void DotNetExecute(this ICakeContext context, FilePath assemblyPath, ProcessArgumentBuilder arguments)
+        {
+            context.DotNetExecute(assemblyPath, arguments, null);
+        }
+
+        /// <summary>
+        /// Execute an assembly with arguments in the specific path with settings.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="assemblyPath">The assembly path.</param>
+        /// <param name="arguments">The arguments.</param>
+        /// <param name="settings">The settings.</param>
+        /// <example>
+        /// <code>
+        /// var settings = new DotNetExecuteSettings
+        /// {
+        ///     FrameworkVersion = "1.0.3"
+        /// };
+        ///
+        /// DotNetExecute("./bin/Debug/app.dll", "--arg", settings);
+        /// </code>
+        /// </example>
+        [CakeMethodAlias]
+        [CakeAliasCategory("Execute")]
+        [CakeNamespaceImport("Cake.Common.Tools.DotNet.Execute")]
+        public static void DotNetExecute(this ICakeContext context, FilePath assemblyPath, ProcessArgumentBuilder arguments, DotNetExecuteSettings settings)
+        {
+            if (context is null)
+            {
+                throw new ArgumentNullException(nameof(context));
+            }
+
+            if (assemblyPath is null)
+            {
+                throw new ArgumentNullException(nameof(assemblyPath));
+            }
+
+            if (settings is null)
+            {
+                settings = new DotNetExecuteSettings();
+            }
+
+            var executor = new DotNetExecutor(context.FileSystem, context.Environment, context.ProcessRunner, context.Tools);
+            executor.Execute(assemblyPath, arguments, settings);
+        }
+    }
+}
